First pass at corrections window
This just is as couple of basic corrections
around capitalization.

The main feature here is the way that I communicate with the
open window.
Instead of starting a slow TCP server,
I use multiprocessing pipes.
and instead of starting a new process every time I show the window,
I reuse the existing window and just hide it
when complete.

I've tested as many recent error conditions as I can think of
it seems to be fairly stable.
But it is still very much a beta!
I have not tested on Python 2, and I'm sure it will break.
I'm not sure if Pysides even works on Python 2.
The multi- processing module should be functional,
but there will definitely be bugs.
The Python dev team did a major re-factor
of the multi-processing module in Python 3.4.
But it should at least be tested so that it imports
on Python 2 without crashing.